% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Greeting Function}
\usage{
hello(name)
}
\arguments{
\item{name}{participant name.}
}
\value{
Greeting of participant.
}
\description{
This function grrets the conference participant.
}
\note{
This function is a bit boring but that is ok.
}
\examples{
hello("Dorothee")
}
