% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{weather}
\alias{weather}
\title{My weather Function}
\usage{
weather()
}
\value{
Forecast.
}
\description{
This function provides a random weather forecast.
}
\note{
Take this weather forecast with a grain of salt.
}
\examples{
weather()
}
\seealso{
Check Google for a more correct forecast.
}
